Refactor functions in ``args_handling.py``
Restructure the implementation of several functions in ``args_handling.py`` by extracting them into separate functions now residing in ``configs_handling.py``.

- Reorganized code for improved modularity
- Created dedicated functions in ``configs_handling.py``
- Adjusted imports and references accordingly
